Create CustomESPWiFiManager_KenTaylorSwitchFS.ino
Im using the esp01s on a usb programmer connected to my laptop, no dht22 device.  The programmer hast GPIO0 connected to GND via a push button which is used to put the esp01s into programming mode and its what i use to call the portal as well.  The programmer also has 3.3V and EN pin connected to each other.

Problem is that i can configure the esp01s in portal mode and save the ssid/pwd and apikey but after EXIT PORTAL it tries to reconnect to the SSID on its own but fails with connection result 0.